chore: polish release
