+++ internal/collector/iis/iis_test.go
@@ -40,7 +40,7 @@ func TestIISDeduplication(t *testing.T) {
 	// Map key and Name should be identical, as there is no suffix starting with "#"
 	expected["bar_2"] = perflibAPP_POOL_WAS{Name: "bar_2", Frequency_Object: 5}
 
-	deduplicated := dedupIISNames(start)
+	deduplicated := deduplicateIISNames(start)
 	if !reflect.DeepEqual(expected, deduplicated) {
 		t.Errorf("Flattened values do not match!\nExpected result: %+v\nActual result: %+v", expected, deduplicated)
 	}
